+++ oap-server/server-query-plugin/promql-plugin/src/main/java/org/apache/skywalking/oap/query/promql/rt/PromQLExprQueryVisitor.java
@@ -460,9 +460,15 @@ public class PromQLExprQueryVisitor extends PromQLParserBaseVisitor<ParseResult>
             scalarResult.setResultType(ParseResultType.SCALAR);
             return scalarResult;
         } else if (left.getResultType() == ParseResultType.METRICS_RANGE && right.getResultType() == ParseResultType.SCALAR) {
-            return matrixScalarBinaryOp((MetricsRangeResult) left, (ScalarResult) right, opType);
+            return matrixScalarBinaryOp(
+                (MetricsRangeResult) left,
+                rangeValue -> scalarBinaryOp(rangeValue, ((ScalarResult) right).getValue(), opType)
+            );
         } else if (left.getResultType() == ParseResultType.SCALAR && right.getResultType() == ParseResultType.METRICS_RANGE) {
-            return matrixScalarBinaryOp((MetricsRangeResult) right, (ScalarResult) left, opType);
+            return matrixScalarBinaryOp(
+                (MetricsRangeResult) right,
+                rangeValue -> scalarBinaryOp(((ScalarResult) left).getValue(), rangeValue, opType)
+            );
         } else if (left.getResultType() == ParseResultType.METRICS_RANGE && right.getResultType() == ParseResultType.METRICS_RANGE) {
             try {
                 return matrixBinaryOp((MetricsRangeResult) left, (MetricsRangeResult) right, opType);